+++ framework/Database/BlueprintInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Framework\Database;
+
+interface BlueprintInterface
+{
+    public function build(): array;
+}